feat: WishList 레포지토리 분리 - Jpa 레포지토리/순수 Java 레포지토리 분리
